+++ rumqttc/examples/tls.rs
@@ -10,9 +10,9 @@ async fn main() -> Result<(), Box<dyn Error>> {
     pretty_env_logger::init();
     color_backtrace::install();
 
-    let mut mqtt_options = MqttOptions::new("test-1", "mqtt.example.server", 8883);
-    mqtt_options.set_keep_alive(std::time::Duration::from_secs(5));
-    mqtt_options.set_credentials("username", "password");
+    let mut mqttoptions = MqttOptions::new("test-1", "mqtt.example.server", 8883);
+    mqttoptions.set_keep_alive(std::time::Duration::from_secs(5));
+    mqttoptions.set_credentials("username", "password");
 
     // Use rustls-native-certs to load root certificates from the operating system.
     let mut root_cert_store = rustls::RootCertStore::empty();
@@ -25,14 +25,14 @@ async fn main() -> Result<(), Box<dyn Error>> {
         .with_root_certificates(root_cert_store)
         .with_no_client_auth();
 
-    mqtt_options.set_transport(Transport::tls_with_config(client_config.into()));
+    mqttoptions.set_transport(Transport::tls_with_config(client_config.into()));
 
-    let (_client, mut eventloop) = AsyncClient::new(mqtt_options, 10);
+    let (_client, mut eventloop) = AsyncClient::new(mqttoptions, 10);
 
     loop {
         match eventloop.poll().await {
             Ok(Event::Incoming(Incoming::Publish(p))) => {
-                println!("Topic: {}, Payload: {:?}", p.topic, p.payload)
+                println!("Topic: {}, Payload: {:?}", p.topic, p.payload);
             }
             Ok(Event::Incoming(i)) => {
                 println!("Incoming = {:?}", i);