+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+----------------
+rumqttc
+-----------
+- Change variants in `ClientError` to not expose dependence on flume/`SendError` (#420)
+- Revert erroring out when Subscription filter list is empty (#422)
+- Remove the ability to cancel `EventLoop.poll()` mid execution (#421)
+
+-----------
+
+### R13
+----------------
+rumqttc v0.13.0
+-----------
+- Add code in `rumqttc::v5`, moving towards support for operating the client with MQTT 5 (#351, #393, #398)
+- Add missing `self.inflight += 1` in `MqttState.save_pubrel()` (#389)
+- Error out when Subscription filter list is empty (#392)
+- Make public `rumqttc::client::Iter` (#402)
+
+-----------
+
+### R12
+----------------
+rumqttc v0.12.0
+-----------
+- Enable compilation without `rustls` as a dependency using `--no-default-features` (#365)
+- Rework variants of `ConnectionError` (#370)
+- New constructor `MqttOptions::parse()` using [`url`](https://docs.rs/url) (#379)
+- Use `get_mut()` instead of index based access to ensure no panic (#384)
+- Better error messages (#385) 
+
+rumqttd v0.11.0
+-----------
+- Enable compilation without rustls as a dependency using `--no-default-features` (#365)
+- Better error messages (#385)
+-----------
+
 ### R11
 ----------------
 rumqttc v0.11.0